# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Clone API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterRequest {
  username: string;
  password: string;
}

export interface LoginRequest {
  username: string;
  password: string;
}

export interface User {
  id: number;
  username: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface FriendRequest {
  id: number;
  senderId: number;
  receiverId: number;
  status: string;
  sender?: User;
  receiver?: User;
}

export interface FriendRequestSend {
  receiverId: number;
}

export interface FriendRequestAccept {
  requestId: number;
}

export interface Message {
  id: number;
  content: string;
  senderId: number;
  senderUsername: string;
  createdAt: string;
  dmUserId?: number | null;
  groupId?: number | null;
}

export interface Group {
  id: number;
  name: string;
  createdById: number;
  members?: User[];
}

export interface CreateGroupRequest {
  name: string;
  memberIds: number[];
}

export type SearchUsersParams = {
  q: string;
};