# lib/api-zod/src/generated/types/friendRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Clone API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface FriendRequest {
  id: number;
  senderId: number;
  receiverId: number;
  status: string;
  sender?: User;
  receiver?: User;
}