# lib/api-zod/src/generated/types/group.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Clone API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from "./user";

export interface Group {
  id: number;
  name: string;
  createdById: number;
  members?: User[];
}