# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Clone API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  content: string;
  senderId: number;
  senderUsername: string;
  createdAt: string;
  dmUserId?: number | null;
  groupId?: number | null;
}